requirements.txt
Flask
flask-cors
python-dotenv
requests
openai
langchain
langchain-community
langchain-openai
pytest